2020 day 13 solved

+++ 2019/Day13.cs
@@ -74,7 +74,7 @@ namespace _2019
                 if (item.Key.Y>row)
                 {
                     text += Environment.NewLine;
-                    row = item.Key.Y;
+                    row = (int)item.Key.Y;
                 }
                 switch (item.Value)
                 {
